// spring-ai-alibaba-admin/Dify-DSL-Import-Fidelity-Patch/files/frontend/packages/main/src/types/appManage.ts
export interface IHistoryConfig {
  history_switch: boolean;
  history_max_round: number;
}

export interface IGlobalVariableItem {
  key: string;
  type: IValueType;
  desc: string;
  default_value: string | number | boolean | object | null;
}

export interface IWorkFlowConfig {
  nodes: IBizNode[];
  edges: IBizEdge[];
  global_config: {
    history_config: IHistoryConfig;
    variable_config: {
      conversation_params: Array<IGlobalVariableItem>;
    };
  };
}